requirements.txt
pyvi==0.1
torch~=1.8.1
transformers==4.5.0
emot==2.1
flask==2.0.1
beautifulsoup4==4.9.3
selenium==3.141.0
requests
PyYAML
tqdm
pandas==1.3.4
matplotlib==3.5.0
